+++ src/index.js
@@ -11,7 +11,7 @@ export { default as ControlledSwitch } from './components/ControlledSwitch';
 export { default as UncontrolledSwitch } from './components/UncontrolledSwitch';
 export { Credentials } from './components/Authz/Credentials';
 export { CellContentLoader, ChartLoader } from './components/ContentLoader';
-export { DateTime } from './components/DateTime';
+export { DateTime } from './components/DateTime/DateTime';
 export { RelativeDateTime } from './components/DateTime/RelativeDateTime.jsx';
 export { default as timeToString } from './components/DateTime/timeToString';
 export { default as useDateFNSLocale } from './components/DateTime/useDateFNSLocale';
